requirements.txt
autopep8==2.0.0
azure-core==1.26.1
azure-cosmos==4.3.0
azure-functions==1.12.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
charset-normalizer==2.1.1
idna==3.4
pycodestyle==2.9.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
